Move primitive-only support into stdlib/prims; drop dead caml_do_read
