+++ runtime/io.c
@@ -220,31 +220,6 @@ CAMLexport void caml_close_channel(struct channel *channel)
   caml_stat_free(channel);
 }
 
-CAMLexport file_offset caml_channel_size(struct channel *channel)
-{
-  file_offset here, end;
-  int fd;
-
-  caml_channel_check_pending(channel);
-  /* We extract data from [channel] before dropping the OCaml lock, in case
-     someone else touches the block. */
-  fd = channel->fd;
-  here = channel->flags & CHANNEL_TEXT_MODE ? -1 : channel->offset;
-  caml_enter_blocking_section_no_pending();
-  if (here == -1) {
-    here = lseek(fd, 0, SEEK_CUR);
-    if (here == -1) goto error;
-  }
-  end = lseek(fd, 0, SEEK_END);
-  if (end == -1) goto error;
-  if (lseek(fd, here, SEEK_SET) != here) goto error;
-  caml_leave_blocking_section();
-  return end;
- error:
-  caml_leave_blocking_section();
-  caml_sys_error(NO_ARG);
-}
-
 CAMLexport int caml_channel_binary_mode(struct channel *channel)
 {
   return channel->flags & CHANNEL_TEXT_MODE ? 0 : 1;
@@ -373,35 +348,7 @@ CAMLexport void caml_really_putblock(struct channel *channel,
   }
 }
 
-CAMLexport void caml_seek_out(struct channel *channel, file_offset dest)
-{
-  file_offset res;
-  caml_flush(channel);
-  caml_enter_blocking_section_no_pending();
-  res = lseek(channel->fd, dest, SEEK_SET);
-  if (res < 0 || res != dest) {
-    caml_leave_blocking_section();
-    caml_sys_error(NO_ARG);
-  }
-  caml_leave_blocking_section();
-  channel->offset = dest;
-}
-
-CAMLexport file_offset caml_pos_out(struct channel *channel)
-{
-  return channel->offset + (file_offset)(channel->curr - channel->buff);
-}
-
 /* Input */
-
-int caml_do_read(int fd, char *p, unsigned int n)
-{
-  int r;
-  do {
-    r = caml_read_fd(fd, 0, p, n);
-  } while (r == -1 && errno == EINTR);
-  return r;
-}
 
 CAMLexport unsigned char caml_refill(struct channel *channel)
 {
@@ -486,76 +433,6 @@ CAMLexport intnat caml_really_getblock(struct channel *chan, char *p, intnat n)
     k -= r;
   }
   return n - k;
-}
-
-CAMLexport void caml_seek_in(struct channel *channel, file_offset dest)
-{
-  file_offset res;
-  if (dest >= channel->offset - (channel->max - channel->buff)
-      && dest <= channel->offset
-      && (channel->flags & CHANNEL_TEXT_MODE) == 0) {
-    channel->curr = channel->max - (channel->offset - dest);
-  } else {
-    caml_enter_blocking_section_no_pending();
-    res = lseek(channel->fd, dest, SEEK_SET);
-    if (res < 0 || res != dest) {
-      caml_leave_blocking_section();
-      caml_sys_error(NO_ARG);
-    }
-    caml_leave_blocking_section();
-    channel->offset = dest;
-    channel->curr = channel->max = channel->buff;
-  }
-}
-
-CAMLexport file_offset caml_pos_in(struct channel *channel)
-{
-  return channel->offset - (file_offset)(channel->max - channel->curr);
-}
-
-intnat caml_input_scan_line(struct channel *channel)
-{
-  char * p;
-  int n;
- again:
-  caml_channel_check_pending(channel);
-  p = channel->curr;
-  do {
-    if (p >= channel->max) {
-      /* No more characters available in the buffer */
-      if (channel->curr > channel->buff) {
-        /* Try to make some room in the buffer by shifting the unread
-           portion at the beginning */
-        memmove(channel->buff, channel->curr, channel->max - channel->curr);
-        n = channel->curr - channel->buff;
-        channel->curr -= n;
-        channel->max -= n;
-        p -= n;
-      }
-      if (channel->max >= channel->end) {
-        /* Buffer is full, no room to read more characters from the input.
-           Return the number of characters in the buffer, with negative
-           sign to indicate that no newline was encountered. */
-        return -(channel->max - channel->curr);
-      }
-      /* Fill the buffer as much as possible */
-      n = caml_read_fd(channel->fd, channel->flags,
-                       channel->max, channel->end - channel->max);
-      if (n == -1) {
-        if (errno == EINTR) goto again; else caml_sys_io_error(NO_ARG);
-      }
-      else if (n == 0) {
-        /* End-of-file encountered. Return the number of characters in the
-           buffer, with negative sign since we haven't encountered
-           a newline. */
-        return -(channel->max - channel->curr);
-      }
-      channel->offset += n;
-      channel->max += n;
-    }
-  } while (*p++ != '\n');
-  /* Found a newline. Return the length of the line, newline included. */
-  return (p - channel->curr);
 }
 
 /* OCaml entry points for the I/O functions.  Wrap struct channel *
